refactor(utils): add rogu/utils folder
