+++ src/rogu/utils/index.ts
@@ -1,2 +1,3 @@
+export * from "./assignmentMaterial";
 export * from "./date";
 export * from "./message";
